refact: move cli code to index
